+++ R/construct_buffer.R
@@ -6,9 +6,10 @@
 #' @param shape String specifying plot shape. Either "circle" or "square"
 #' @param size Size of sample plot. Equals the radius for circles or the
 #' side-length for squares in mapunits
+#' @param return_sp If true, SpatialPolygons are returned.
 #'
 #' @return
-#' SpatialPolygons
+#' matrix or SpatialPolygons
 #'
 #' @examples
 #' coords <- matrix(c(10, 5, 25, 15, 5, 25), ncol = 2, byrow = TRUE)
@@ -20,49 +21,53 @@
 #' @keywords internal
 #'
 #' @export
-construct_buffer <- function(coords, shape, size) UseMethod("construct_buffer")
+construct_buffer <- function(coords, shape, size, return_sp) UseMethod("construct_buffer")
 
 #' @name construct_buffer
 #' @export
-construct_buffer.SpatialPoints <- function(coords, shape, size) {
+construct_buffer.SpatialPoints <- function(coords, shape, size, return_sp = TRUE) {
 
     coords <- matrix(sp::coordinates(coords), ncol = 2)
-    construct_buffer(coords, shape, size)
+
+    construct_buffer(coords, shape, size, return_sp = return_sp)
 }
 
 #' @name construct_buffer
 #' @export
-construct_buffer.SpatialPointsDataFrame <- function(coords, shape, size) {
+construct_buffer.SpatialPointsDataFrame <- function(coords, shape, size, return_sp = TRUE) {
 
     coords <- matrix(sp::coordinates(coords), ncol = 2)
-    construct_buffer(coords, shape, size)
+
+    construct_buffer(coords, shape, size, return_sp = return_sp)
 }
 
 #' @name construct_buffer
 #' @export
-construct_buffer.MULTIPOINT <- function(coords, shape, size) {
+construct_buffer.MULTIPOINT <- function(coords, shape, size, return_sp = TRUE) {
 
     coords <- matrix(sf::st_coordinates(coords)[, 1:2], ncol = 2)
-    construct_buffer(coords, shape, size)
+
+    construct_buffer(coords, shape, size, return_sp = return_sp)
 }
 
 #' @name construct_buffer
 #' @export
-construct_buffer.POINT <- function(coords, shape, size) {
+construct_buffer.POINT <- function(coords, shape, size, return_sp = TRUE) {
 
     coords <- matrix(sf::st_coordinates(coords)[, 1:2], ncol = 2)
-    construct_buffer(coords, shape, size)
+
+    construct_buffer(coords, shape, size, return_sp = return_sp)
 }
 
 #' @name construct_buffer
 #' @export
-construct_buffer.sf <- function(coords, shape, size) {
+construct_buffer.sf <- function(coords, shape, size, return_sp = TRUE) {
 
     if (all(sf::st_geometry_type(coords) %in% c("POINT", "MULTIPOINT"))) {
 
         coords <- matrix(sf::st_coordinates(coords)[, 1:2], ncol = 2)
 
-        construct_buffer(coords, shape, size)
+        construct_buffer(coords, shape, size, return_sp = return_sp)
     }
 
     else{stop("Only POINT or MULTIPOINT features supported!!11!!1!!")}
@@ -70,13 +75,13 @@ construct_buffer.sf <- function(coords, shape, size) {
 
 #' @name construct_buffer
 #' @export
-construct_buffer.sfc <- function(coords, shape, size) {
+construct_buffer.sfc <- function(coords, shape, size, return_sp = TRUE) {
 
     if (all(sf::st_geometry_type(coords) %in% c("POINT", "MULTIPOINT"))) {
 
         coords <- matrix(sf::st_coordinates(coords)[, 1:2], ncol = 2)
 
-        construct_buffer(coords, shape, size)
+        construct_buffer(coords, shape, size, return_sp = return_sp)
     }
 
     else{stop("Only POINT or MULTIPOINT features supported!!11!!1!!")}
@@ -84,7 +89,7 @@ construct_buffer.sfc <- function(coords, shape, size) {
 
 #' @name construct_buffer
 #' @export
-construct_buffer.matrix <- function(coords, shape , size) {
+construct_buffer.matrix <- function(coords, shape , size, return_sp = TRUE) {
 
     if (shape == "circle") {
 
@@ -94,24 +99,27 @@ construct_buffer.matrix <- function(coords, shape , size) {
         x_circle <- outer(circle_points_x,  coords[, 1], `+`)
         y_circle <- outer(circle_points_y,  coords[, 2], `+`)
 
-        sample_plots_coords <- cbind(matrix(x_circle, ncol = 1),
-                                     matrix(y_circle, ncol = 1),
-                                     rep(seq_len(nrow(coords)), each = 100))
+        sample_plots <- cbind(matrix(x_circle, ncol = 1),
+                              matrix(y_circle, ncol = 1),
+                              rep(seq_len(nrow(coords)), each = 100))
 
-        sample_plots_coords_split <- split(sample_plots_coords[, -3], sample_plots_coords[, 3])
+        if (return_sp) {
 
-        sample_plots <- lapply(X = sample_plots_coords_split, FUN = function(x) {
-            sp::Polygon(cbind(x[1:100], x[101:200]))
-        })
+            sample_plots <- split(sample_plots[, -3], sample_plots[, 3])
 
-        sample_plots <- sp::SpatialPolygons(lapply(X = seq_along(sample_plots), FUN = function(y) {
-            sp::Polygons(list(sample_plots[[y]]), ID = y)
-        }))
+            sample_plots <- lapply(X = sample_plots, FUN = function(x) {
+                sp::Polygon(cbind(x[1:100], x[101:200]))
+            })
+
+            sample_plots <- sp::SpatialPolygons(lapply(X = seq_along(sample_plots), FUN = function(y) {
+                sp::Polygons(list(sample_plots[[y]]), ID = y)
+            }))
+        }
     }
 
     else if (shape == "square") {
 
-        sample_plots_coords <- cbind(
+        sample_plots <- cbind(
             matrix(
                 c(coords[, 1] - size,
                   coords[, 1] - size,
@@ -128,16 +136,17 @@ construct_buffer.matrix <- function(coords, shape , size) {
             rep(seq_len(nrow(coords)), times = 4)
         )
 
-        sample_plots_coords_split <- split(sample_plots_coords[, -3],
-                                           sample_plots_coords[, 3])
+        if (return_sp) {
 
-        sample_plots <- lapply(X = sample_plots_coords_split, FUN = function(x) {
-            sp::Polygon(cbind(x[1:4], x[5:8]))
-        })
+            sample_plots <- split(sample_plots[, -3],
+                                  sample_plots[, 3])
 
-        sample_plots <- sp::SpatialPolygons(lapply(X = seq_along(sample_plots), FUN = function(y) {
-            sp::Polygons(list(sample_plots[[y]]), ID = y)
-        }))
+            sample_plots <- lapply(X = sample_plots, FUN = function(x) {
+                sp::Polygon(cbind(x[1:4], x[5:8]))})
+
+            sample_plots <- sp::SpatialPolygons(lapply(X = seq_along(sample_plots), FUN = function(y) {
+                sp::Polygons(list(sample_plots[[y]]), ID = y)}))
+        }
     }
 
     else{