+++ NEWS.md
@@ -18,6 +18,7 @@
     * `calculate_lsm()` can now print the nlayers progress for RasterStacks, RasterBricks and lists
     * replaced `cat()` with `message()`
     * `calculate_lsm()` returns an error message if selected metrics do not exist
+    * `construct_buffer()` can now return a matrix with coords instead of polygons
 
 # landscapemetrics 1.0
 * New functions